// src/models/employee-model.ts
import * as mongoose from "mongoose";
import { Types } from "mongoose";

interface CommonAttributes {
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber: string;
  gender: string;
  photo: string;
}

export interface DEmployee extends CommonAttributes {}

export interface IEmployee extends CommonAttributes, mongoose.Document {}